Config settings are now deserialized into struct.

+++ server/src/filesystem.rs
@@ -1,10 +1,28 @@
 use crate::config;
 use std::path::PathBuf;
 
+pub fn posts_directory() -> PathBuf {
+    format!("{}/posts", config::get().data_dir).into()
+}
+
+pub fn generated_thumbnails_directory() -> PathBuf {
+    format!("{}/generated-thumbnails", config::get().data_dir).into()
+}
+
+pub fn temporary_upload_directory() -> PathBuf {
+    format!("{}/temporary-uploads", config::get().data_dir).into()
+}
+
+pub fn temporary_upload_filepath(filename: &str) -> PathBuf {
+    format!("{}/temporary-uploads/{}", config::get().data_dir, filename).into()
+}
+
 pub fn purge_temporary_uploads() -> std::io::Result<()> {
-    let data_directory = config::read_required_string("data_dir");
-    let temp_path = PathBuf::from(format!("{data_directory}/temporary-uploads"));
-    for entry in std::fs::read_dir(temp_path)? {
+    let temp_path = temporary_upload_directory();
+    if !temp_path.exists() {
+        return Ok(());
+    }
+    for entry in std::fs::read_dir(temporary_upload_directory())? {
         let path = entry?.path();
         std::fs::remove_file(path)?;
     }
